Matrice upload format

+++ src/Entity/RateCard.php
@@ -39,7 +39,7 @@ class RateCard
     private $priceRateCard;
 
     /**
-     * @ORM\OneToMany(targetEntity="App\Entity\Simulation", mappedBy="ratecard")
+     * @ORM\OneToMany(targetEntity="App\Entity\Simulation", mappedBy="ratecard", orphanRemoval=true)
      */
     private $simulations;
 
@@ -94,12 +94,12 @@ class RateCard
         return $this;
     }
 
-    public function getPriceRateCard(): ?int
+    public function getPriceRateCard(): ?float
     {
         return $this->priceRateCard;
     }
 
-    public function setPriceRateCard(int $priceRateCard): self
+    public function setPriceRateCard(float $priceRateCard): self
     {
         $this->priceRateCard = $priceRateCard;
